+++ ui/app/src/components/users/UserEditorForm.tsx
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Action } from '@perses-dev/core';
-import { getSubmitText, getTitleAction, UserEditorSchemaType, userSchema } from '@perses-dev/plugin-system';
+import { Action, UserEditorSchemaType, UserResource, userSchema } from '@perses-dev/core';
+import { getSubmitText, getTitleAction } from '@perses-dev/plugin-system';
 import React, { DispatchWithoutAction, Fragment, useMemo, useState } from 'react';
 import { Controller, FormProvider, SubmitHandler, useFieldArray, useForm } from 'react-hook-form';
 import { Alert, Box, Button, Divider, FormControl, IconButton, Stack, TextField, Typography } from '@mui/material';
@@ -21,7 +21,6 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import MinusIcon from 'mdi-material-ui/Minus';
 import PlusIcon from 'mdi-material-ui/Plus';
 import DeleteIcon from 'mdi-material-ui/DeleteOutline';
-import { UserResource } from '../../model/user-client';
 import { useIsExternalProviderEnabled, useIsNativeProviderEnabled } from '../../context/Config';
 
 interface UserEditorFormProps {
