+++ ui/core/src/schema/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2023 The Perses Authors
+// Copyright 2024 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -22,4 +22,3 @@ export * from './rolebinding';
 export * from './secret';
 export * from './user';
 export * from './variable';
-export * from './ValidationProvider';
