+++ ui/core/src/schema/metadata.ts
