+++ ui/core/src/schema/plugin.ts
