+++ ui/core/src/schema/rolebinding.ts
@@ -12,7 +12,7 @@
 // limitations under the License.
 
 import { z } from 'zod';
-import { RoleBinding, RoleBindingSpec, Subject } from '@perses-dev/core';
+import { RoleBinding, RoleBindingSpec, Subject } from '../model';
 import { nameSchema, metadataSchema, projectMetadataSchema } from './metadata';
 
 export const subjectSchema: z.ZodSchema<Subject> = z.object({
