+++ ui/core/src/schema/secret.ts
