+++ ui/plugin-system/src/context/index.ts
@@ -12,3 +12,4 @@
 // limitations under the License.
 
 export * from './ProjectStoreProvider';
+export * from './ValidationProvider';
