+++ ui/plugin-system/src/index.ts
@@ -17,5 +17,4 @@ export * from './model';
 export * from './runtime';
 export * from './test-utils';
 export * from './utils';
-export * from './validation';
 export * from './context';
